refactor: remove duplicate code for 'up' and 'cd' commands by creating changeDir function
