+++ commands/fs/cd.js
@@ -1,6 +1,5 @@
-import path from "path";
-import { promises as fs } from "fs";
 import { ERROR_MESSAGES } from "../../constants/fsMessages.js";
+import { changeDirectory } from "../../utils/changeDir.js";
 
 export async function cd(inputArgs) {
   if (inputArgs.length === 0) {
@@ -8,19 +7,5 @@ export async function cd(inputArgs) {
   }
 
   const targetPath = inputArgs.join(" ").replace(/^"|"$/g, "").trim();
-  const absolutePath = path.resolve(process.cwd(), targetPath);
-
-  try {
-    await fs.access(absolutePath);
-
-    const stats = await fs.stat(absolutePath);
-    if (!stats.isDirectory()) {
-      return { success: false, message: ERROR_MESSAGES.OPERATION_FAILED };
-    }
-
-    process.chdir(absolutePath);
-    return { success: true, message: "" };
-  } catch (error) {
-    return { success: false, message: ERROR_MESSAGES.OPERATION_FAILED };
-  }
+  return await changeDirectory(targetPath);
 }
